+++ CHANGES.md
@@ -9,6 +9,7 @@
 * Compiler: codegen: specialize string equality
 * Compiler: codegen: more specialization for %int_add, %int_sub
 * Compiler: recognize and optimize String.concat
+* Compiler: more inlining - duplicate small function.
 
 ## Bug fixes